+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - New #191: Add `ReplaceManagedBotToken` method.
 - New #191: Add `PreparedKeyboardButton` type.
 - New #191: Add `SavePreparedKeyboardButton` method.
+- Chg #191: Replace `correctOptionId` field with `correctOptionIds` in `Poll` type.
 
 ## 0.17 March 1, 2026